Move sentryapps RPC service to the toplevel module
Remove the imports from the module init as they cause cycles.
