+++ src/sentry/api/endpoints/integrations/sentry_apps/interaction.py
@@ -10,7 +10,7 @@ from sentry.api.base import StatsMixin, region_silo_endpoint
 from sentry.api.bases import RegionSentryAppBaseEndpoint, SentryAppStatsPermission
 from sentry.api.bases.sentryapps import COMPONENT_TYPES
 from sentry.models.integrations.sentry_app import SentryApp
-from sentry.services.hybrid_cloud.app import RpcSentryApp, app_service
+from sentry.sentry_apps.services.app import RpcSentryApp, app_service
 from sentry.tsdb.base import TSDBModel
 
 logger = logging.getLogger(__name__)
