+++ src/sentry/api/serializers/models/sentry_app_component.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from sentry.api.serializers import Serializer, register
 from sentry.api.serializers.base import serialize
 from sentry.models.integrations.sentry_app_component import SentryAppComponent
-from sentry.services.hybrid_cloud.app import SentryAppEventDataInterface
+from sentry.sentry_apps.services.app import SentryAppEventDataInterface
 
 
 @register(SentryAppComponent)
