+++ src/sentry/incidents/endpoints/organization_alert_rule_available_action_index.py
@@ -22,7 +22,7 @@ from sentry.incidents.logic import (
 from sentry.incidents.models.alert_rule import AlertRuleTriggerAction
 from sentry.incidents.serializers import ACTION_TARGET_TYPE_TO_STRING
 from sentry.models.organization import Organization
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation, app_service
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation, app_service
 from sentry.services.hybrid_cloud.integration import RpcIntegration
 from sentry.silo.base import region_silo_function
