+++ src/sentry/incidents/endpoints/serializers/alert_rule.py
@@ -27,8 +27,8 @@ from sentry.models.integrations.sentry_app_installation import prepare_ui_compon
 from sentry.models.rule import Rule
 from sentry.models.rulesnooze import RuleSnooze
 from sentry.models.user import User
-from sentry.services.hybrid_cloud.app import app_service
-from sentry.services.hybrid_cloud.app.model import RpcSentryAppComponentContext
+from sentry.sentry_apps.services.app import app_service
+from sentry.sentry_apps.services.app.model import RpcSentryAppComponentContext
 from sentry.services.hybrid_cloud.user import RpcUser
 from sentry.services.hybrid_cloud.user.service import user_service
 from sentry.snuba.models import SnubaQueryEventType
