+++ src/sentry/incidents/utils/sentry_apps.py
@@ -7,7 +7,7 @@ from sentry.auth.access import NoAccess
 from sentry.incidents.logic import get_filtered_actions
 from sentry.incidents.models.alert_rule import AlertRuleTriggerAction
 from sentry.incidents.serializers import AlertRuleTriggerActionSerializer
-from sentry.services.hybrid_cloud.app import app_service
+from sentry.sentry_apps.services.app import app_service
 
 
 def trigger_sentry_app_action_creators_for_incidents(alert_rule_data: Mapping[str, Any]) -> None:
