+++ src/sentry/mediators/external_issues/issue_link_creator.py
@@ -8,7 +8,7 @@ from sentry.mediators.mediator import Mediator
 from sentry.mediators.param import Param
 from sentry.models.group import Group
 from sentry.models.platformexternalissue import PlatformExternalIssue
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation
 from sentry.services.hybrid_cloud.user import RpcUser
