+++ src/sentry/mediators/sentry_app_installations/updater.py
@@ -5,7 +5,7 @@ from sentry.constants import SentryAppInstallationStatus
 from sentry.mediators.mediator import Mediator
 from sentry.mediators.param import Param
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation
 
 
 class Updater(Mediator):
