+++ src/sentry/mediators/token_exchange/grant_exchanger.py
@@ -15,7 +15,7 @@ from sentry.models.apitoken import ApiToken
 from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.models.user import User
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation
 from sentry.silo.safety import unguarded_write
