+++ src/sentry/mediators/token_exchange/refresher.py
@@ -12,7 +12,7 @@ from sentry.models.apitoken import ApiToken
 from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
 from sentry.models.user import User
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation
 
 
 class Refresher(Mediator):
