+++ src/sentry/mediators/token_exchange/validator.py
@@ -7,7 +7,7 @@ from sentry.mediators.param import Param
 from sentry.models.apiapplication import ApiApplication
 from sentry.models.integrations.sentry_app import SentryApp
 from sentry.models.user import User
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation
 
 
 class Validator(Mediator):
