+++ src/sentry/receivers/sentry_apps.py
@@ -12,7 +12,7 @@ from sentry.models.project import Project
 from sentry.models.team import Team
 from sentry.models.user import User
 from sentry.sentry_apps.apps import consolidate_events
-from sentry.services.hybrid_cloud.app import RpcSentryAppInstallation, app_service
+from sentry.sentry_apps.services.app import RpcSentryAppInstallation, app_service
 from sentry.services.hybrid_cloud.user import RpcUser
 from sentry.signals import (
     comment_created,
