+++ src/sentry/rules/actions/notify_event_service.py
@@ -17,7 +17,7 @@ from sentry.plugins.base import plugins
 from sentry.rules.actions.base import EventAction
 from sentry.rules.actions.services import PluginService
 from sentry.rules.base import CallbackFuture
-from sentry.services.hybrid_cloud.app import RpcSentryAppService, app_service
+from sentry.sentry_apps.services.app import RpcSentryAppService, app_service
 from sentry.services.hybrid_cloud.integration import integration_service
 from sentry.services.hybrid_cloud.organization.serial import serialize_rpc_organization
 from sentry.tasks.sentry_apps import notify_sentry_app
