+++ src/sentry/rules/actions/sentry_apps/utils.py
@@ -6,7 +6,7 @@ from typing import Any
 from rest_framework import serializers
 
 from sentry.constants import SENTRY_APP_ACTIONS
-from sentry.services.hybrid_cloud.app import app_service
+from sentry.sentry_apps.services.app import app_service
 
 
 def trigger_sentry_app_action_creators_for_issues(
