+++ src/sentry/sentry_apps/components.py
@@ -11,8 +11,8 @@ from django.utils.http import urlencode
 from sentry.mediators.external_requests.select_requester import SelectRequester
 from sentry.models.integrations.sentry_app_component import SentryAppComponent
 from sentry.models.integrations.sentry_app_installation import SentryAppInstallation
-from sentry.services.hybrid_cloud.app.model import RpcSentryAppComponent, RpcSentryAppInstallation
-from sentry.services.hybrid_cloud.app.serial import serialize_sentry_app_installation
+from sentry.sentry_apps.services.app.model import RpcSentryAppComponent, RpcSentryAppInstallation
+from sentry.sentry_apps.services.app.serial import serialize_sentry_app_installation
 from sentry.utils import json
