+++ src/sentry/sentry_apps/services/app/__init__.py
@@ -1,0 +1,2 @@
+from .model import *  # noqa
+from .service import *  # noqa
