+++ src/sentry/services/hybrid_cloud/app/__init__.py
@@ -1,2 +1,8 @@
-from .model import *  # noqa
-from .service import *  # noqa
+# TODO(hybridcloud) Remove once getsentry usage is updated
+from sentry.sentry_apps.services.model import RpcSentryAppInstallation
+from sentry.sentry_apps.services.service import app_service
+
+__all__ = (
+    "RpcSentryAppInstallation",
+    "app_service",
+)
