+++ tests/sentry/incidents/endpoints/test_organization_alert_rule_available_action_index.py
@@ -9,7 +9,7 @@ from sentry.incidents.models.alert_rule import AlertRuleTriggerAction
 from sentry.integrations.pagerduty.utils import add_service
 from sentry.models.integrations import SentryAppComponent, SentryAppInstallation
 from sentry.models.integrations.organization_integration import OrganizationIntegration
-from sentry.services.hybrid_cloud.app.serial import serialize_sentry_app_installation
+from sentry.sentry_apps.services.app.serial import serialize_sentry_app_installation
 from sentry.services.hybrid_cloud.integration.serial import serialize_integration
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import APITestCase
