+++ tests/sentry/mediators/external_issues/test_creator.py
@@ -1,6 +1,6 @@
 from sentry.mediators.external_issues.creator import Creator
 from sentry.models.platformexternalissue import PlatformExternalIssue
-from sentry.services.hybrid_cloud.app import app_service
+from sentry.sentry_apps.services.app import app_service
 from sentry.testutils.cases import TestCase
