+++ tests/sentry/mediators/external_issues/test_issue_link_creator.py
@@ -4,7 +4,7 @@ import responses
 from sentry.coreapi import APIUnauthorized
 from sentry.mediators.external_issues.issue_link_creator import IssueLinkCreator
 from sentry.models.platformexternalissue import PlatformExternalIssue
-from sentry.services.hybrid_cloud.app import app_service
+from sentry.sentry_apps.services.app import app_service
 from sentry.services.hybrid_cloud.user.serial import serialize_rpc_user
 from sentry.testutils.cases import TestCase
