+++ tests/sentry/mediators/external_requests/test_select_requester.py
@@ -3,7 +3,7 @@ import responses
 
 from sentry.coreapi import APIError
 from sentry.mediators.external_requests.select_requester import SelectRequester
-from sentry.services.hybrid_cloud.app import app_service
+from sentry.sentry_apps.services.app import app_service
 from sentry.testutils.cases import TestCase
 from sentry.utils.sentry_apps import SentryAppWebhookRequestsBuffer
