+++ tests/sentry/sentry_apps/test_sentry_app_component_preparer.py
@@ -2,7 +2,7 @@ from unittest.mock import call, patch
 
 from sentry.models.organization import Organization
 from sentry.sentry_apps.components import SentryAppComponentPreparer
-from sentry.services.hybrid_cloud.app.serial import serialize_sentry_app_installation
+from sentry.sentry_apps.services.app.serial import serialize_sentry_app_installation
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
